Token events service data getter was moved from event service to specified token events service.
