+++ apps/web-api/src/event/token-event.service.ts
@@ -1,10 +1,16 @@
 import { ITokenEntities, Tokens } from '@entities/Tokens';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { In, Not, Repository, SelectQueryBuilder } from 'typeorm';
 import { nesting } from '@unique-nft/utils/address';
-import { EventMethod } from '@common/constants';
+import { EventMethod, EventSection } from '@common/constants';
 import { TokenEventDTO } from './token-event.dto';
+import { IDataListResponse, IGQLQueryArgs } from '../utils/gql-query-args';
+import { GraphQLResolveInfo } from 'graphql';
+import { BaseService } from '../utils/base.service';
+import { Event } from '@entities/Event';
+import { EventDTO } from './event.dto';
+import { IRelations } from '../utils/base.service.types';
 
 type ITokenInfo = Pick<
   Tokens,
@@ -27,9 +33,48 @@ interface IEventTransferValues {
   collectionId: number;
 }
 
+const EXTRINSIC_RELATION_ALIAS = 'Extrinsic';
+
+const aliasFields = {
+  action: 'method',
+  author: 'signer',
+  result: 'success',
+};
+
+const relationsFields = {
+  author: EXTRINSIC_RELATION_ALIAS,
+  result: EXTRINSIC_RELATION_ALIAS,
+  fee: EXTRINSIC_RELATION_ALIAS,
+};
+
+const customQueryFields = {
+  collection_id: '("Event".data::json->>0)::int',
+  token_id: '("Event".data::json->>1)::int',
+};
+
 @Injectable()
-export class TokenEventService {
-  constructor(@InjectRepository(Tokens) private repo: Repository<Tokens>) {}
+export class TokenEventService extends BaseService<Event, EventDTO> {
+  constructor(
+    @InjectRepository(Tokens) private tokenRepo: Repository<Tokens>,
+    @InjectRepository(Event) private repo: Repository<Event>,
+  ) {
+    super({ aliasFields, relationsFields, customQueryFields });
+  }
+
+  protected getConditionField(
+    qb: SelectQueryBuilder<Event>,
+    field: string,
+  ): string {
+    if (field === 'collection_id') {
+      return `("${qb.alias}".data::json->>0)::int`;
+    }
+
+    if (field === 'token_id') {
+      return `("${qb.alias}".data::json->>1)::int`;
+    }
+
+    return super.getConditionField(qb, field);
+  }
 
   public async mapEventData(events: TokenEventDTO[]) {
     for (const event of events) {
@@ -62,7 +107,7 @@ export class TokenEventService {
   }
 
   private async findTokens(args: ITokenEntities[]): Promise<ITokenInfo[]> {
-    const qb = this.repo.createQueryBuilder();
+    const qb = this.tokenRepo.createQueryBuilder();
     qb.select(['token_id', 'collection_id::int', 'token_name', 'image']);
     qb.where(args);
 
@@ -75,5 +120,39 @@ export class TokenEventService {
     } catch {
       return null;
     }
+  }
+
+  public async find(
+    queryArgs: IGQLQueryArgs<Partial<TokenEventDTO>>,
+    queryInfo: GraphQLResolveInfo,
+  ): Promise<IDataListResponse<TokenEventDTO>> {
+    const qb = this.repo.createQueryBuilder();
+
+    const queryFields = this.getQueryFields(queryInfo);
+
+    const relations = {
+      [EXTRINSIC_RELATION_ALIAS]: {
+        table: 'extrinsic',
+        on: `"${EXTRINSIC_RELATION_ALIAS}".block_index = "Event".block_index`,
+      },
+    } as IRelations;
+
+    this.applySelect(qb, queryArgs, queryFields, relations);
+
+    qb.where({
+      phase: Not('Initialization'),
+      section: EventSection.COMMON,
+      method: In([
+        EventMethod.ITEM_CREATED,
+        EventMethod.TRANSFER,
+        EventMethod.ITEM_DESTROYED,
+      ]),
+    });
+
+    this.applyLimitOffset(qb, queryArgs);
+    this.applyWhereCondition(qb, queryArgs);
+    this.applyOrderCondition(qb, queryArgs);
+
+    return this.getDataAndCount(qb, queryArgs);
   }
 }
